requirements.txt
python-kasa == 0.9.1
prometheus-client == 0.21.1